+++ src/interfaces/migrator.interface.ts
@@ -1,0 +1,5 @@
+export interface IMigrator<TOld, TNew> {
+  get oldVersion(): string;
+  get newVersion(): string;
+  migrate(old: TOld): TNew;
+}