+++ src/services/vscode-service.ts
@@ -1,7 +1,34 @@
 import * as vscode from 'vscode';
+import { TrackingStatus } from '../enums/tracking-status';
 import { IInteractionService } from '../interfaces/interaction-service.interface';
 
 export class VSCodeInteractionService implements IInteractionService {
+  statusBarButton: vscode.StatusBarItem;
+
+  constructor() {
+    this.statusBarButton = vscode.window.createStatusBarItem('id', vscode.StatusBarAlignment.Left);
+  }
+
+  setButtonStatus(status: TrackingStatus): void {
+    switch (status) {
+      case TrackingStatus.NotTracking:
+        this.statusBarButton.text = '▶ Start Tracking';
+        this.statusBarButton.tooltip = 'Click to start tracking time';
+        this.statusBarButton.command = 'time-tracker.startTrackingTime';
+        this.statusBarButton.show();
+        break;
+      case TrackingStatus.Tracking:
+        this.statusBarButton.text = 'Tracking...';
+        this.statusBarButton.tooltip = 'Click to stop tracking time';
+        this.statusBarButton.command = 'time-tracker.stopTrackingTime';
+        this.statusBarButton.show();
+        break;
+      default:
+        this.statusBarButton.hide();
+        break;
+    }
+  }
+
   getWorkspacePath(): string {
     try {
       const result = vscode.workspace.workspaceFolders?.[0].uri.fsPath;
@@ -20,7 +47,6 @@ export class VSCodeInteractionService implements IInteractionService {
   async showInputBox(message: string, placeholder: string | undefined = undefined, suggestions: string[] = []): Promise<string | undefined> {
     let disposable: vscode.Disposable | undefined;
     if (suggestions.length > 0) {
-      // TODO: try
       disposable = vscode.languages.registerCompletionItemProvider('*', {
         provideCompletionItems(document, position) {
           let completionItems = suggestions.map((suggestion) => {