+++ CHANGELOG.md
@@ -5,6 +5,13 @@
     Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 -->
 
+## [0.1.1]
+
+### Added
+
+- StatusBar Button
+- `times.json`-Migrations
+
 ## [0.1.0]
 
 ### Added